pyproject.toml
[tool.poetry]
name = "iot-bot"
version = "0.1.0"
description = ""
authors = ["Rate980"]
readme = "README.md"
packages = [{ include = "iot_bot" }]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.2.3"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
pynacl = "^1.5.0"
aiofiles = "^23.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
